requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.5.0
paddleocr==2.7.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Pillow==10.0.0
Werkzeug==3.0.0
